OE-234 Remove CardCreator logic from OE
