+++ Simplified.xcodeproj/xcshareddata/xcschemes/Open eBooks.xcscheme
@@ -82,9 +82,6 @@
             </BuildableReference>
             <SkippedTests>
                <Test
-                  Identifier = "NYPLSignInBusinessLogicTests/testCardCreatorSupport()">
-               </Test>
-               <Test
                   Identifier = "NYPLSignInBusinessLogicTests/testUpdateUserAccountWithSAMLAuthentication()">
                </Test>
             </SkippedTests>
